+++ composeml/tests/test_label_times.py
@@ -1,56 +1,11 @@
 import pandas as pd
-import pytest
-
-from ..label_times import LabelTimes
-
-
-@pytest.fixture
-def labels():
-    records = [
-        {
-            'label_id': 0,
-            'customer_id': 1,
-            'time': '2014-01-01 00:45:00',
-            'my_labeling_function': 226.92999999999998
-        },
-        {
-            'label_id': 1,
-            'customer_id': 1,
-            'time': '2014-01-01 00:48:00',
-            'my_labeling_function': 47.95
-        },
-        {
-            'label_id': 2,
-            'customer_id': 2,
-            'time': '2014-01-01 00:01:00',
-            'my_labeling_function': 283.46000000000004
-        },
-        {
-            'label_id': 3,
-            'customer_id': 2,
-            'time': '2014-01-01 00:04:00',
-            'my_labeling_function': 31.54
-        },
-    ]
-
-    dtype = {'time': 'datetime64[ns]'}
-    df = pd.DataFrame.from_records(records).astype(dtype)
-
-    df = df.set_index('label_id')
-    df = df[['customer_id', 'time', 'my_labeling_function']]
-
-    labels = LabelTimes(df)
-    labels.settings = {
-        'name': 'my_labeling_function',
-        'target_entity': 'customer_id',
-    }
-
-    return labels
 
 
 def test_threshold(labels):
     given_labels = labels.threshold(200)
     answer = [True, False, True, False]
+
+    labels = labels.copy()
     labels['my_labeling_function'] = answer
 
     pd.testing.assert_frame_equal(given_labels, labels)
@@ -84,6 +39,7 @@ def test_bins(labels):
         pd.Interval(31.288, 157.5, closed='right'),
     ]
 
+    labels = labels.copy()
     labels['my_labeling_function'] = pd.Categorical(answer, ordered=True)
     pd.testing.assert_frame_equal(given_labels, labels)
 
@@ -98,21 +54,10 @@ def test_quantile_bins(labels):
         pd.Interval(31.538999999999998, 137.44, closed='right'),
     ]
 
+    labels = labels.copy()
     labels['my_labeling_function'] = pd.Categorical(answer, ordered=True)
     pd.testing.assert_frame_equal(given_labels, labels)
 
 
 def test_describe(labels):
     assert labels.describe() is None
-
-
-def test_distribution_plot(labels):
-    labels = labels.threshold(200)
-    plot = labels.plot.distribution()
-    assert plot.get_title() == 'label_distribution'
-
-
-def test_count_by_time_plot(labels):
-    labels = labels.threshold(200)
-    plot = labels.plot.count_by_time()
-    assert plot.get_title() == 'count_by_time'
